+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added general plotter for geometry objects and data structures based on the artist registration mechanism.
 * Added support for multimesh files to OBJ reader/writer.
 * Added support for attaching and detaching meshes in `compas.robots.RobotModelArtist` and drawing them.
+* Added basic implementation of `compas.datastructures.Assembly`.
 
 ### Changed